Move calculate_status to separate module
